Create AssembleTopoWXWeatherData.R
This script had to be run in chunks due to timeouts and RAM used since we are aggregating over a wide area for daily data.  After all the data are assembled, we get mean values per day for Tmin and Tmax and then we average to a daily mean temp.  These are all eventually bound into a single file and saved to disk, representing daily data over 1948-2016 for one "location" (e.g. grid cell)